+++ src/main/java/chess/domain/game/ChessGame.java
@@ -1,8 +1,8 @@
 package chess.domain.game;
 
 import chess.domain.board.Board;
-import chess.domain.board.Position;
 import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 
 public class ChessGame {
