+++ src/test/java/chess/domain/PositionFixture.java
@@ -1,20 +1,20 @@
 package chess.domain;
 
-import static chess.domain.board.FileCoordinate.A;
-import static chess.domain.board.FileCoordinate.B;
-import static chess.domain.board.FileCoordinate.C;
-import static chess.domain.board.FileCoordinate.D;
-import static chess.domain.board.FileCoordinate.E;
-import static chess.domain.board.FileCoordinate.F;
-import static chess.domain.board.RankCoordinate.FIVE;
-import static chess.domain.board.RankCoordinate.FOUR;
-import static chess.domain.board.RankCoordinate.ONE;
-import static chess.domain.board.RankCoordinate.SEVEN;
-import static chess.domain.board.RankCoordinate.SIX;
-import static chess.domain.board.RankCoordinate.THREE;
-import static chess.domain.board.RankCoordinate.TWO;
+import static chess.domain.position.FileCoordinate.A;
+import static chess.domain.position.FileCoordinate.B;
+import static chess.domain.position.FileCoordinate.C;
+import static chess.domain.position.FileCoordinate.D;
+import static chess.domain.position.FileCoordinate.E;
+import static chess.domain.position.FileCoordinate.F;
+import static chess.domain.position.RankCoordinate.FIVE;
+import static chess.domain.position.RankCoordinate.FOUR;
+import static chess.domain.position.RankCoordinate.ONE;
+import static chess.domain.position.RankCoordinate.SEVEN;
+import static chess.domain.position.RankCoordinate.SIX;
+import static chess.domain.position.RankCoordinate.THREE;
+import static chess.domain.position.RankCoordinate.TWO;
 
-import chess.domain.board.Position;
+import chess.domain.position.Position;
 
 public class PositionFixture {
     public static final Position C_4 = new Position(C, FOUR);
